Using timestamp to determine epoches instead of block number 1. changed all relevant logic of the epoch to use timestamp 2. upgraded buidler version and rely on its features to do timestamp testing 3. fixed all relevant tests

+++ contractsSol5/mock/MockDAO.sol
@@ -52,17 +52,11 @@ contract MockDAO is IKyberDAO, Utils4 {
         feeBps = _feeBps;
     }
 
-    function getLatestNetworkFeeData() 
-        external 
-        view 
-        returns(uint256, uint256) {
+    function getLatestNetworkFeeData() external view returns(uint, uint) {
         return (feeBps, expiryTimestamp);
     }
 
-    function getLatestNetworkFeeDataWithCache() 
-        external 
-        returns(uint256, uint256) 
-    {
+    function getLatestNetworkFeeDataWithCache() external returns(uint, uint) {
         data++;
         return (feeBps, expiryTimestamp);
     }
